lean-learn-backend/requirements.txt
fastapi
uvicorn
python-dotenv
# sqlalchemy==1.4.31
pymongo[srv]
python-dotenv
certifi
google
google-generativeai
langchain-openai
langsmith